+++ src/app/products/components/product-list/product-list.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import { Subscription } from 'rxjs';
+import { Observable, of, Subscription } from 'rxjs';
 import { CartService } from 'src/app/cart/services/cart.service';
 import { Categories } from '../../enums/categories.enum';
 import { Product } from '../../../shared/models/product.model';
@@ -11,7 +11,19 @@ import { ProductsService } from '../../services/products.service';
   styleUrls: ['./product-list.component.scss']
 })
 export class ProductListComponent implements OnInit {
-  products: Product[] = [];
+  products$: Observable<Product[]> = of([]);
+
+  sortOptions = [{
+    name: 'name',
+    value: 'name'
+  }, {
+    name: 'price',
+    value: 'price'
+  }];
+
+  selectedOption = this.sortOptions[0].name;
+
+  order = false;
 
   private productsSub!: Subscription;
 
@@ -21,10 +33,7 @@ export class ProductListComponent implements OnInit {
   ) { }
 
   ngOnInit(): void {
-    this.productsSub = this.productsService.getProducts()
-      .subscribe((products: Product[]) => {
-        this.products = products;
-      });
+    this.products$ = this.productsService.getProducts();
   }
 
   addToCard(product: Product): void {