+++ src/app/shared/shared.module.ts
@@ -3,16 +3,27 @@ import { CommonModule } from '@angular/common';
 import { ProductInfoComponent } from './components';
 import { HighlightDirective } from './directives/highlight.directive';
 import { HostClickDirective } from './directives/host-click.directive';
+import { OrderByPipe } from './pipes/order-by.pipe';
+import { FormsModule } from '@angular/forms';
 
 @NgModule({
   declarations: [
     ProductInfoComponent,
     HighlightDirective,
-    HostClickDirective
+    HostClickDirective,
+    OrderByPipe
   ],
   imports: [
-    CommonModule
+    CommonModule,
+    FormsModule
   ],
-  exports: [ProductInfoComponent, HighlightDirective, HostClickDirective]
+  exports: [
+    CommonModule,
+    FormsModule,
+    ProductInfoComponent,
+    HighlightDirective,
+    HostClickDirective,
+    OrderByPipe
+  ]
 })
 export class SharedModule { }