+++ changelog.md
@@ -4,6 +4,33 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.0] - 2021-05-23
+
+## Added
+- create order by directive, use for products list
+
+## Updated
+- update products service to return observable
+- added async pipe to product list and cart list components
+- add uppercase pipe to product info component
+- update shared module
+
+## [0.3.0] - 2021-05-23
+
+## Added
+- ConfigOptionsService
+- ConstantsService
+- GeneratorService
+- genID
+- LocalStorageService
+- host click directive
+
+## Updated
+- First component to try all aded services
+- CartService
+
+
+
 ## [0.2.0] - 2021-05-23
 
 ### Added